# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sallehly AI Video Engine API
  version: "1"
  description: >
    Public/internal API for apps/api - implemented and tested end-to-end
    (fastapi.testclient.TestClient) as of Phase 4. Every path below is a
    real route in apps/api/src/api/routes/. Route handlers call
    state.orchestrator (an IProjectOrchestrator), never
    CreativeDirector/CreativeCompiler/GenerationPipeline directly - see
    docs/adr/0010-persistence-and-lifecycle.md. Business schemas are
    referenced directly from packages/schemas/json/ so this contract
    cannot drift from the internal pipeline contracts.
servers:
  - url: https://api.sallehly.ai
    description: Production (future)
  - url: http://localhost:8000
    description: Local dev

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: >
        Creates a user and its personal default workspace
        (workspace_id=ws_<user_id>). See packages/auth - there is no
        team/workspace-membership model yet, every user gets exactly one
        workspace that is theirs alone.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "201":
          description: The created user (never includes password_hash/password_salt)
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/user.schema.json"
        "409":
          description: Email already registered

  /auth/login:
    post:
      summary: Authenticate and receive a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Bearer token + the authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    $ref: "../../packages/schemas/json/user.schema.json"
        "401":
          description: Invalid email or password

  /users/me:
    get:
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/user.schema.json"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /projects:
    get:
      summary: List projects in the caller's own workspace
      description: >
        workspace_id is always the authenticated user's own
        (current_user.workspace_id) - never taken from a query parameter,
        so a token can only ever list its own workspace's projects.
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "../../packages/schemas/json/project.schema.json"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a project from a creative brief
      description: >
        Creates the Project (status=created). Does not trigger planning -
        call POST /projects/{id}/generate-plan next. workspace_id and
        created_by are derived from the bearer token, not accepted from
        the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, target_duration_sec, aspect_ratio]
              properties:
                prompt: { type: string }
                target_duration_sec: { type: number, exclusiveMinimum: 0 }
                aspect_ratio: { type: string, enum: ["16:9", "9:16", "1:1", "21:9", "4:5"] }
                reference_asset_ids:
                  type: array
                  items: { type: string }
                style_preset_id: { type: string }
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /projects/{project_id}:
    get:
      summary: Get a project's current state
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /projects/{project_id}/generate-plan:
    post:
      summary: Run Creative Planning (AI Director + Creative Compiler through storyboard generation)
      description: >
        created -> planning -> waiting_storyboard_approval. Internally:
        CreativeDirector.generate_director_plan then
        CreativeCompiler.compile_storyboard.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project now waiting_storyboard_approval
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/approve-storyboard:
    post:
      summary: Approve the storyboard (gate 1), compiling the render plan
      description: >
        Requires status=waiting_storyboard_approval.
        waiting_storyboard_approval -> waiting_render_approval. Internally:
        CreativeCompiler.approve_storyboard then
        CreativeCompiler.compile_render_plan.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project now waiting_render_approval
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/reject-storyboard:
    post:
      summary: Reject the storyboard with feedback; regenerates it
      description: >
        Requires status=waiting_storyboard_approval. Records feedback,
        then immediately re-invokes CreativeDirector.regenerate_with_feedback
        and CreativeCompiler.compile_storyboard, leaving the project
        waiting_storyboard_approval again with a revised storyboard.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feedback]
              properties:
                feedback:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Project waiting_storyboard_approval again, with a revised storyboard
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/approve-render:
    post:
      summary: Approve the render plan (gate 2)
      description: Requires status=waiting_render_approval. waiting_render_approval -> approved.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project now approved, ready for generate-video
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/reject-render:
    post:
      summary: Reject the render plan with feedback; recompiles it
      description: >
        Requires status=waiting_render_approval. Records feedback, then
        recompiles the render plan (optionally at a different
        quality_tier), leaving the project waiting_render_approval again.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feedback]
              properties:
                feedback:
                  type: array
                  items: { type: string }
                quality_tier:
                  type: string
                  enum: [preview, standard, final]
      responses:
        "200":
          description: Project waiting_render_approval again, with a recompiled render plan
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/generate-video:
    post:
      summary: Submit every shot in the approved render plan to the Video Engine
      description: >
        Requires status=approved. approved -> generating -> completed (or
        failed, if any shot's GenerationJob failed). Internally:
        GenerationPipeline.generate_plan, one GenerationJob per shot.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project now completed or failed; generation_job_ids/asset_ids populated
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/retry-generation:
    post:
      summary: Retry generation after a failed attempt
      description: >
        Requires status=failed. Re-runs GenerationPipeline.generate_plan
        against the same already-approved render plan (no re-approval
        needed - approval covers the creative content, not the GPU job
        outcome). failed -> generating -> completed (or failed again).
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project now completed or failed again
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/project.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"

  /projects/{project_id}/assets:
    get:
      summary: List the project's generated assets
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id: { type: string }
                  assets:
                    type: array
                    items:
                      $ref: "../../packages/schemas/json/asset_record.schema.json"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /assets/upload:
    post:
      summary: Upload a reference asset (e.g. an image)
      description: >
        Frontend flow: upload here to get an asset_id, then pass it in
        POST /projects' reference_asset_ids. project_id is optional - a
        reference image is commonly uploaded before its project exists,
        in which case the asset is filed under the caller's workspace.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                project_id: { type: string }
      responses:
        "201":
          description: Registered asset
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/asset_record.schema.json"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: project_id was given but doesn't exist

  /jobs/{job_id}:
    get:
      summary: Get a generation job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: GenerationJob
          content:
            application/json:
              schema:
                $ref: "../../packages/schemas/json/generation_job.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"

  /jobs/{job_id}/status:
    get:
      summary: Get a generation job's status only (lighter-weight poll target)
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  status: { type: string, enum: [queued, running, completed, failed] }
                  retry_count: { type: integer }
                  error_message: { type: ["string", "null"] }
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: { type: string }
    JobId:
      name: job_id
      in: path
      required: true
      schema: { type: string }

  responses:
    Unauthorized:
      description: Missing/malformed/invalid or expired bearer token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    Forbidden:
      description: Authenticated, but not the owner of this project
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    InvalidTransition:
      description: >
        The project's current status doesn't allow this transition (e.g.
        approving a storyboard that isn't waiting_storyboard_approval).
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

security:
  - BearerAuth: []